// simulator/src/generated/furnace/server-message.ts
export enum ServerMessage {
  NONE = 0,
  State = 1,
  Ack = 2,
  HistoryResponse = 3,
  ProgramListResponse = 4,
  ProgramContentResponse = 5,
  PreferencesResponse = 6,
  DebugInfoResponse = 7,
  LogListResponse = 8,
  LogContentResponse = 9,
  Error = 10
}